% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{adjust_alpha_optimize}
\alias{adjust_alpha_optimize}
\title{Adjust the alpha level of the envelope test using optimization}
\source{
https://github.com/TeemuSailynoja/simultaneous-confidence-bands
}
\usage{
adjust_alpha_optimize(alpha, N, K)
}
\description{
Adjust the alpha level of the envelope test using optimization
}
