% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{constrain_interval}
\alias{constrain_interval}
\title{Constrain vector to be inside interval}
\usage{
constrain_interval(x, lower = 0, upper = 1, verbose = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{lower}{The smallest possible value of `x`.}

\item{upper}{The largest possible value of `x`.}
}
\value{
A vector `x` where values outside the interval are moved inside.
}
\description{
Constrain vector to be inside interval
}
