% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{empirical_coverage}
\alias{empirical_coverage}
\title{Calculates the empirical coverage of a vector `x`.}
\usage{
empirical_coverage(x, nominal_coverage)
}
\description{
Calculates the empirical coverage of a vector `x`.
}
