% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{get_lims}
\alias{get_lims}
\title{This function simulates upper and lower limits `lims$lower` and `lims$upper` such that:
\deqn{P({lims$lower[i] <= sum(unif(n) < 1/i) <= lims$upper[i]} for any i = 1, \ldots, K - 1) = 1 - alpha}}
\source{
https://github.com/TeemuSailynoja/simultaneous-confidence-bands
}
\usage{
get_lims(alpha, n, K)
}
\arguments{
\item{alpha}{Confidence level between 0 and 1.}

\item{n}{Sample size.}

\item{K}{Granularity of the uniform partition of the unit interval.}
}
\value{
List containing the upper and lower simultaneous confidence bands
evaluated at \eqn{z_i = i / K} for \eqn{i = 1, \ldots, K - 1}.
}
\description{
For more information see "Graphical Test for Discrete Uniformity and its Applications in
Goodness of Fit Evaluation and Multiple Sample Comparison" (Säilynoja, Bürkner, Vehtari).
}
