% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-idx.R
\name{mutate_dummy}
\alias{mutate_dummy}
\title{Mutate a new set of columns for dummy variable coding of `var`.}
\usage{
mutate_dummy(df, var)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{A categorical variable in `df`.}
}
\value{
A dataframe with additional columns for the dummy variables.
}
\description{
The `var` column is a categorical variable which may be expressed in terms of
`length(unique(df$var))` columns with levels zero or one using `model.matrix`.
}
