% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{p_interior}
\alias{p_interior}
\title{Probability of the ECDF being completely within the envelope as z2}
\source{
https://github.com/TeemuSailynoja/simultaneous-confidence-bands
}
\usage{
p_interior(p_int, x1, x2, z1, z2, gamma, N)
}
\description{
Probability of the ECDF being completely within the envelope as z2
}
