% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-idx.R
\name{repeat_matrix}
\alias{repeat_matrix}
\title{Repeat a matrix along the diagonal.}
\usage{
repeat_matrix(M, n)
}
\arguments{
\item{M}{A matrix.}

\item{n}{The number of times to be repeated.}
}
\value{
A repeated matrix.
}
\description{
The output should have dimensions `dim(M) * n`.
}
