% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{run_commit_push}
\alias{run_commit_push}
\title{Run and optionally commit or push a collection of reports.}
\usage{
run_commit_push(reports, commit = TRUE, push = FALSE)
}
\arguments{
\item{reports}{A vector of report names.}

\item{commit}{Should the reports be commited? Defaults to `TRUE`.}

\item{push}{Should the reports be pushed to remote? Defaults to `FALSE`.}
}
\description{
Run and optionally commit or push a collection of reports.
}
