% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{stable_softmax}
\alias{stable_softmax}
\title{Compute the numerically-stable softmax of a vector.}
\usage{
stable_softmax(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The softmax of `x`
}
\description{
Compute the numerically-stable softmax of a vector.
}
