% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-prior.R
\name{tau_fixed}
\alias{tau_fixed}
\title{Fixing the precision prior in `R-INLA`.}
\usage{
tau_fixed(x)
}
\arguments{
\item{x}{Used to specify the `initial`.}
}
\value{
A prior that can be passed to `R-INLA`.
}
\description{
Fixing the precision prior in `R-INLA`.
}
