% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-prior.R
\name{tau_pc}
\alias{tau_pc}
\title{Penalised complexity precision prior for `R-INLA`.}
\usage{
tau_pc(x, u, alpha)
}
\arguments{
\item{x}{Used to specify the `initial`.}

\item{u}{Upper threshold.}

\item{alpha}{Probability that the standard deviation exceeds
the upper threshold.}
}
\value{
A prior that can be passed to `R-INLA`.
}
\description{
Penalised complexity precision prior for `R-INLA`.
}
