% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-idx.R
\name{to_int}
\alias{to_int}
\title{Categorical to indicators.}
\usage{
to_int(x)
}
\arguments{
\item{x}{A categorical column.}
}
\value{
Indicator variables column.
}
\description{
Categorical to indicators.
}
